requirements.txt
Flask==2.0.2
azure-storage-blob
pandas